Разделил функции по файлам
