+++ js/modules/validation/popup.js
@@ -1,0 +1,81 @@
+const initializeErrorElement = () => {
+  const errorTemplate = document.querySelector('#error').content.querySelector('.error');
+  const errorElement = errorTemplate.cloneNode(true);
+  errorElement.classList.add('visually-hidden');
+  const errorMessageElement = errorElement.querySelector('.error__message');
+  const errorInfoElement = document.createElement('p');
+  errorInfoElement.classList.add('error__info');
+  errorMessageElement.append(errorInfoElement);
+
+  const errorButtonElement = errorElement.querySelector('.error__button');
+  errorButtonElement.addEventListener('click', () => {
+    errorElement.classList.add('visually-hidden');
+  });
+  document.body.append(errorElement);
+  return errorElement;
+};
+
+export const showError = (text) => {
+  let errorElement = document.querySelector('.error');
+  if (!errorElement) {
+    errorElement = initializeErrorElement();
+  }
+  errorElement.querySelector('.error__info').textContent = text;
+
+  const OnEscKeydown = (evt) => {
+    if (evt.key === 'Escape') {
+      errorElement.classList.add('visually-hidden');
+    }
+  };
+  document.body.addEventListener('keydown', OnEscKeydown, { once: true });
+
+  const OnDocumentClick = (evt) => {
+    for (const node of errorElement.children) {
+      if (node === evt.target) {
+        return;
+      }
+    }
+    errorElement.classList.add('visually-hidden');
+    document.removeEventListener('click', OnDocumentClick);
+  };
+  document.addEventListener('click', OnDocumentClick);
+
+  errorElement.classList.remove('visually-hidden');
+
+};
+
+const initializeSuccessElement = () => {
+  const successTemplate = document.querySelector('#success').content.querySelector('.success');
+  const successElement = successTemplate.cloneNode(true);
+  successElement.classList.add('visually-hidden');
+  document.body.append(successElement);
+  return successElement;
+};
+
+export const showSuccess = () => {
+  let successElement = document.querySelector('.success');
+  if (!successElement) {
+    successElement = initializeSuccessElement();
+  }
+
+  const OnEscKeydown = (evt) => {
+    if (evt.key === 'Escape') {
+      successElement.classList.add('visually-hidden');
+    }
+  };
+  document.body.addEventListener('keydown', OnEscKeydown, { once: true });
+
+  const OnDocumentClick = (evt) => {
+    for (const node of successElement.children) {
+      if (node === evt.target) {
+        return;
+      }
+    }
+    successElement.classList.add('visually-hidden');
+    document.removeEventListener('click', OnDocumentClick);
+  };
+  document.addEventListener('click', OnDocumentClick);
+
+  successElement.classList.remove('visually-hidden');
+
+};
